finishing master

+++ app/Http/Controllers/ResortController.php
@@ -5,8 +5,10 @@ namespace App\Http\Controllers;
 
 
 use App\Helper\CustomController;
+use App\Helper\Formula;
 use App\Models\Resort;
 use App\Models\ServiceUnit;
+use Illuminate\Support\Facades\Validator;
 
 class ResortController extends CustomController
 {
@@ -26,6 +28,11 @@ class ResortController extends CustomController
 
     public function index($service_unit_id)
     {
+        $hasAccessServiceUnit = $this->hasServiceUnitAccess($service_unit_id);
+        if (!$hasAccessServiceUnit) {
+            abort(403);
+        }
+        $access = $this->getRoleAccess(Formula::APPMasterResort);
         $service_unit = ServiceUnit::findOrFail($service_unit_id);
         if ($this->request->ajax()) {
             $data = Resort::with(['service_unit:id,name'])
@@ -34,15 +41,39 @@ class ResortController extends CustomController
             return $this->basicDataTables($data);
         }
         return view('admin.master-data.resort.index')->with([
-            'service_unit' => $service_unit
+            'service_unit' => $service_unit,
+            'access' => $access,
         ]);
     }
 
+    private $rule = [
+        'name' => 'required',
+    ];
+
+    private $message = [
+        'name.required' => 'kolom nama resort wajib di isi',
+    ];
+
     public function store($service_unit_id)
     {
+        //filtering service unit
+        $hasAccessServiceUnit = $this->hasServiceUnitAccess($service_unit_id);
+        if (!$hasAccessServiceUnit) {
+            abort(403);
+        }
+
+        $access = $this->getRoleAccess(Formula::APPMasterResort);
+        if (!$access['is_granted_create']) {
+            abort(403);
+        }
+
         $service_unit = ServiceUnit::findOrFail($service_unit_id);
         if ($this->request->method() === 'POST') {
             try {
+                $validator = Validator::make($this->request->all(), $this->rule, $this->message);
+                if ($validator->fails()) {
+                    return redirect()->back()->withErrors($validator)->with('validator', 'Harap Mengisi Kolom Dengan Benar');
+                }
                 $data_request = [
                     'service_unit_id' => $service_unit_id,
                     'name' => $this->postField('name'),
@@ -58,10 +89,25 @@ class ResortController extends CustomController
 
     public function patch($service_unit_id, $id)
     {
+        //filtering service unit
+        $hasAccessServiceUnit = $this->hasServiceUnitAccess($service_unit_id);
+        if (!$hasAccessServiceUnit) {
+            abort(403);
+        }
+
+        $access = $this->getRoleAccess(Formula::APPMasterResort);
+        if (!$access['is_granted_update']) {
+            abort(403);
+        }
+
         $service_unit = ServiceUnit::findOrFail($service_unit_id);
         $data = Resort::with(['service_unit'])->findOrFail($id);
         if ($this->request->method() === 'POST') {
             try {
+                $validator = Validator::make($this->request->all(), $this->rule, $this->message);
+                if ($validator->fails()) {
+                    return redirect()->back()->withErrors($validator)->with('validator', 'Harap Mengisi Kolom Dengan Benar');
+                }
                 $data_request = [
                     'service_unit_id' => $service_unit_id,
                     'name' => $this->postField('name'),
@@ -80,6 +126,17 @@ class ResortController extends CustomController
 
     public function destroy($service_unit_id, $id)
     {
+        //filtering service unit
+        $hasAccessServiceUnit = $this->hasServiceUnitAccess($service_unit_id);
+        if (!$hasAccessServiceUnit) {
+            abort(403);
+        }
+
+        $access = $this->getRoleAccess(Formula::APPMasterResort);
+        if (!$access['is_granted_delete']) {
+            return $this->jsonErrorResponse('cannot access delete perform...');
+        }
+
         try {
             Resort::destroy($id);
             return $this->jsonSuccessResponse('success');
